Activity 1 - Particle Movement
Particle Activity

In this activity sphere is set to change its movement in Vector3.

A camera script describes if the left click on the mouse is pressed so the camera could be child of the sphere object, following its movement. If the click is not pressed, sets the camera to the original position.